Workaround for nick change issue.
For some reason the exception thrown when a rename fails due to an
in-use nickname is not being caught even though there it's in
a try..catch block!  something to do with flow-js?

For now, I set a flag in the flow instead and quickly skip through
remaining steps of the flow.  It works fine but it looks ugly.